requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
boto3==1.34.111
botocore==1.34.111
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2.12
django-cors-headers==4.3.1
django-seed==0.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
Faker==25.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.8
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.31.0
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
toposort==1.10
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
zope.interface==6.4.post2
